% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{myboot2}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{Number of iterations the function uses}

\item{x}{The sample we will bootstrap}

\item{fun}{The statistic we want to calculate}

\item{alpha}{Determines the value of the confidence interval}

\item{cx}{Helps place the text}

\item{...}{Any other stipulations}
}
\value{
A histogram of the bootstrapped values and ci, and an invsible list
}
\description{
myboot2
}
\examples{
myboot2(x=rnorm(20,mean=10,sd=4), fun="mean")
}
