% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltu.R
\name{mycltu}
\alias{mycltu}
\title{mycltu}
\usage{
mycltu(n, iter, a = 0, b = 10)
}
\arguments{
\item{n}{number of samples}

\item{iter}{number of iterations}

\item{a}{the a value}

\item{b}{the b value}
}
\value{
A plot
}
\description{
mycltu
}
\examples{
mycltu(10,1000)
}
