% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myf.R
\name{myf}
\alias{myf}
\title{Piecewise Function}
\usage{
myf(x, xk, coef)
}
\arguments{
\item{x}{X value, as a vector}

\item{xk}{Xk value, the palce where the pieceline line changes}

\item{coef}{Coefficients of the piecewise line equation, as a vector}
}
\value{
Values as a vector in a piecewise line regression regression equation
}
\description{
Piecewise Function
}
\examples{
\dontrun{myf(c(1:21), 17, quad.lm$coef)}
}
