% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlik.R
\name{mymaxlik}
\alias{mymaxlik}
\title{Maximum Likelihood}
\usage{
mymaxlik(lfun, x, param, ...)
}
\arguments{
\item{lfun}{Log function}

\item{x}{sample}

\item{param}{Parameter being calculated}

\item{...}{Any other needs}
}
\value{
A contoured plot and list
}
\description{
Maximum Likelihood
}
\examples{
\dontrun{mymaxlik(lfun=logbin,x=y, param=seq(0,1, length=1000))}
}
