% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{myncurve}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{Mean of the normal distribution}

\item{sigma}{Standard deviation of the normal distribution}

\item{a}{Numerical value that y is less than}
}
\value{
Graph of probability X <= a in the normal distribution, the probability
}
\description{
myncurve
}
\examples{
myncurve(10,5,2)
}
