% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mynewsample.R
\name{mynewsample}
\alias{mynewsample}
\title{Multiple iterations of a sample 1-10}
\usage{
mynewsample(n, iter = 10, time = 0.5)
}
\arguments{
\item{n}{Size of the population}

\item{iter}{How many times the sample is run}

\item{time}{Time betwene the generation of each plot}
}
\value{
Barplots for each iteration of the random sample
}
\description{
Multiple iterations of a sample 1-10
}
\examples{
\dontrun{mynewsample(n=1000,iter=30,time=1)}
}
