% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{ntickets}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{The number of tickets sold on the flight}

\item{gamma}{The probability of overbooking}

\item{p}{The probability a given person shows up}
}
\value{
A named list of nd,nc,N,,gamma, p
}
\description{
ntickets
}
\examples{
ntickets(200,.02,.95)
}
